Fixed out of bounds error in R_ClipLineToLight.
The third argument for R_ClipLineToLight is an array of length four.
Within the body of the function it is treated as an array of length
six. The only time the function is called it is passed an array of
length six. The definition therefore must be mistaken.

This commit corrects the third argument to be an array of length
six.